requirements.txt
ollama
pandas
pdfminer.six
PyYAML
streamlit